feat: Add crew executions and steps handling to crews handler
