+++ src/database/database-do.ts
@@ -212,43 +212,6 @@ export class DatabaseDO extends DurableObject<Env> {
 				}
 			}
 
-			if (endpoint === '/crews/executions') {
-				const address = url.searchParams.get('address');
-				if (!address) {
-					return createApiResponse('Missing address parameter', 400);
-				}
-				const executions = await getCrewExecutions(this.orm, address);
-				return createApiResponse({
-					message: 'Successfully retrieved crew executions',
-					data: { executions },
-				});
-			}
-
-			if (endpoint === '/crews/executions/add') {
-				if (request.method !== 'POST') {
-					return createApiResponse('Method not allowed', 405);
-				}
-
-				type AddCrewExecutionRequest = {
-					address: string;
-					crewId: number;
-					conversationId: number;
-					input: string;
-				};
-
-				const body: AddCrewExecutionRequest = await request.json();
-				const { address, crewId, conversationId, input } = body;
-
-				if (!address || !crewId || !conversationId || !input) {
-					return createApiResponse('Missing required parameters: address, crewId, conversationId, input', 400);
-				}
-
-				const execution = await addCrewExecution(this.orm, address, crewId, conversationId, input);
-				return createApiResponse({
-					message: 'Successfully created crew execution',
-					data: { execution },
-				});
-			}
 
 			// Pass off to crons handler
 			if (endpoint.startsWith('/crons')) {
@@ -477,103 +440,6 @@ export class DatabaseDO extends DurableObject<Env> {
 			});
 		}
 
-		// Crew execution steps endpoints
-		if (endpoint === '/crews/steps/get') {
-			const executionId = url.searchParams.get('executionId');
-			if (!executionId) {
-				return createApiResponse('Missing executionId parameter', 400);
-			}
-
-			// Get the session token from Authorization header
-			const authHeader = request.headers.get('Authorization');
-			if (!authHeader) {
-				return createApiResponse('Missing authorization header', 401);
-			}
-
-			// Extract token from Bearer format
-			const token = authHeader.replace('Bearer ', '');
-
-			// Verify the token matches the requested address
-			const tokenAddress = await validateSessionToken(this.env, token);
-			if (!tokenAddress.success) {
-				return createApiResponse('Unauthorized access', 403);
-			}
-
-			const steps = await getExecutionSteps(this.orm, parseInt(executionId));
-			return createApiResponse({
-				message: 'Successfully retrieved execution steps',
-				data: { steps },
-			});
-		}
-
-		if (endpoint === '/crews/steps/create') {
-			if (request.method !== 'POST') {
-				return createApiResponse('Method not allowed', 405);
-			}
-
-			// Get the session token from Authorization header
-			const authHeader = request.headers.get('Authorization');
-			if (!authHeader) {
-				return createApiResponse('Missing authorization header', 401);
-			}
-
-			// Extract token from Bearer format
-			const token = authHeader.replace('Bearer ', '');
-
-			// Verify the token
-			const tokenAddress = await validateSessionToken(this.env, token);
-			if (!tokenAddress) {
-				return createApiResponse('Unauthorized access', 403);
-			}
-
-			const stepData = (await request.json()) as UserCrewExecutionStepsTable;
-			if (!stepData.profile_id || !stepData.crew_id || !stepData.execution_id || !stepData.step_type || !stepData.step_data) {
-				return createApiResponse('Missing required fields: profile_id, crew_id, cron_enabled', 400);
-			}
-
-			// Verify the profile_id matches the token address
-			if (stepData.profile_id !== tokenAddress.address) {
-				return createApiResponse('Unauthorized: profile_id does not match token', 403);
-			}
-
-			const step = await createExecutionStep(this.orm, stepData);
-			return createApiResponse({
-				message: 'Successfully created execution step',
-				data: { step },
-			});
-		}
-
-		if (endpoint === '/crews/steps/delete') {
-			if (request.method !== 'DELETE') {
-				return createApiResponse('Method not allowed', 405);
-			}
-
-			const executionId = url.searchParams.get('executionId');
-			if (!executionId) {
-				return createApiResponse('Missing executionId parameter', 400);
-			}
-
-			// Get the session token from Authorization header
-			const authHeader = request.headers.get('Authorization');
-			if (!authHeader) {
-				return createApiResponse('Missing authorization header', 401);
-			}
-
-			// Extract token from Bearer format
-			const token = authHeader.replace('Bearer ', '');
-
-			// Verify the token
-			const tokenAddress = await validateSessionToken(this.env, token);
-			if (!tokenAddress) {
-				return createApiResponse('Unauthorized access', 403);
-			}
-
-			const result = await deleteExecutionSteps(this.orm, parseInt(executionId));
-			return createApiResponse({
-				message: 'Successfully deleted execution steps',
-				data: { result },
-			});
-		}
 
 		return createUnsupportedEndpointResponse(endpoint, this.SUPPORTED_ENDPOINTS);
 	}
